Refactor: pages 구조 정리, style lint 적용
