+++ src/components/Card/SupportCard/SupportCard.module.scss
@@ -2,10 +2,10 @@
 
 .supportCard {
   .content {
+    padding-bottom: 130%;
     overflow: hidden;
     position: relative;
     width: 100%;
-    padding-bottom: 130%;
     border-radius: 8px;
     @media screen and (min-width: $tablet) {
       padding-bottom: 104%;
@@ -18,9 +18,9 @@
     }
     img {
       position: absolute;
-      inset: 0;
       width: 100%;
       height: 100%;
+      inset: 0;
       object-fit: cover;
     }
     button {
