+++ src/pages/MyPage/MyPage.jsx
@@ -1,4 +1,3 @@
-import styles from "pages/MyPage.module.scss";
 import Navbar from "components/Navbar";
 import Layout from "components/Layout/Layout";
 import Button from "components/Button";
@@ -6,6 +5,7 @@ import RoundImage from "components/Image/RoundImage/RoundImage";
 import { useState } from "react";
 import icoPlus from "assets/icons/plus.svg";
 import mockData from "mockData/femaleData.json";
+import styles from "./MyPage.module.scss";
 
 export default function MyPage() {
   const [idols, setIdols] = useState(mockData);
