+++ src/pages/MyPage/MyPage.module.scss
@@ -93,7 +93,7 @@
       &:checked + .labelItem::after {
         position: absolute;
         background:
-          url("../assets/icons/Checkmark.svg") no-repeat center,
+          url("../../assets/icons/Checkmark.svg") no-repeat center,
           linear-gradient(
             271.36deg,
             rgba(249, 110, 104, 0.5) -9.84%,
